# Fix and simplify CalcStartandCount() for BOX rearranger