For test_CalcStartandCount(), update expected result of an
existing test case, and add more test cases (some are from
actual E3SM production runs).

The existing CalcStartandCount() function fails on the updated
test cases, generating correct but suboptimal partitions.

Fix two bugs found in CalcStartandCount() so that it can pass all
updated test cases with expected partitions.

Rewrite CalcStartandCount() as part of code refactoring. The
new version uses less temporary variables and removes an
inefficient while loop for convergence.

Fixes #318