@dqwu 2020-07-02
The following ne1024 F case is run on Cori to verify the performance improvement.
compset: FSCREAM-HR-DYAMOND1
resolution: ne1024pg2_r0125_oRRS18to6v3
knl nodes: 1536
mpi tasks: 12288
mpi tasks per node: 8
PIO_REARRANGER: BOX
PIO_BUFFER_SIZE_LIMIT: 128 MB
PIO_STRIDE: 8 (1536 IO tasks)
lustre stripe count: 248
lustre stripe size: 2 MB

[Load balancing improvement]
3 x 1440 x 2880: use_io_procs increased from 507 to 1536
2 x 3 x 1440 x 2880: use_io_procs increased from 252 to 1536
5 x 3 x 1440 x 2880: use_io_procs increased from 90 to 1530
17 x 1440 x 2880: use_io_procs increased from 68 to 1530
10 x 1440 x 2880: use_io_procs increased from 140 to 1530
15 x 1440 x 2880: use_io_procs increased from 75 to 1530
9 x 1440 x 2880: use_io_procs increased from 162 to 1530
129 x 100663296: use_io_procs increased from 129 to 1419
129 x 25165824: use_io_procs increased from 129 to 1419

[Run time improvement]
Case run time: reduced from 01:06:37 to 00:54:14
CalcStartandCount time: reduced from 0.191 sec to 0.03 sec
PIOc_read_darray time: reduced from 445.467 sec to 135.991 sec
flush_output_buffer time: reduced from 721.913 sec to 620.95 sec
pio_swapm time: reduced from 360.326 sec to 194.42 sec