@dqwu 2020-07-07
> This looks really good. Can you also include stats on overall I/O write throughput ?

ne1024 F case on Cori, size of total output nc files: 3689.268082492 GB

[Old version of CalcStartandCount(), PIO_BUFFER_SIZE_LIMIT = 128 MB]
Case run time: 01:06:37
PIOc_createfile (subtotal): 7.236 sec
PIOc_put_att_tc (subtotal): 0.853 sec
PIOc_put_vars_tc (subtotal): 1.508 sec
PIOc_write_darray (subtotal): 207.723 sec
PIOc_sync (subtotal): 365.291 sec
PIOc_closefile_write_mode (subtotal): 231.426 sec
Total:  814.037 sec
Overall I/O write throughput: 4.532064369 GB / sec

[New version of CalcStartandCount(), PIO_BUFFER_SIZE_LIMIT = 128 MB]
Case run time: 00:54:14
PIOc_createfile (subtotal): 4.189 sec
PIOc_put_att_tc (subtotal): 0.738 sec
PIOc_put_vars_tc (subtotal): 1.825 sec
PIOc_write_darray (subtotal): 221.003 sec
PIOc_sync (subtotal): 248.867 sec
PIOc_closefile_write_mode (subtotal): 213.639 sec
Total: 690.261 sec
Overall I/O write throughput: 5.34474363 GB/sec

[Old version of CalcStartandCount(), PIO_BUFFER_SIZE_LIMIT = 512 MB]
Case run time: 00:56:53
PIOc_createfile (subtotal): 8.5 sec
PIOc_put_att_tc (subtotal): 0.698 sec
PIOc_put_vars_tc (subtotal): 2.424 sec
PIOc_write_darray (subtotal): 4.986 sec
PIOc_sync (subtotal): 627.919 sec
PIOc_closefile_write_mode (subtotal): 235.75 sec
Total:  880.277 sec
Overall I/O write throughput: 4.191030872 GB / sec

[New version of CalcStartandCount(), PIO_BUFFER_SIZE_LIMIT = 512 MB]
Case run time: 00:42:39
PIOc_createfile (subtotal): 6.429 sec
PIOc_put_att_tc (subtotal): 0.576 sec
PIOc_put_vars_tc (subtotal): 1.35 sec
PIOc_write_darray (subtotal): 4.037 sec
PIOc_sync (subtotal): 473.253 sec
PIOc_closefile_write_mode (subtotal): 212.793 sec
Total: 698.438 sec
Overall I/O write throughput: 5.28216976 GB/sec