@dqwu 2020-07-10
The following LND I case is run on Summit, size of total output nc files: 351.505904861 GB
compset: I1850GSWCNPRDCTCBC
resolution: hcru_hcru
summit nodes: 16
mpi tasks: 1344
mpi tasks per node: 84
PIO_REARRANGER: BOX
PIO_STRIDE: 21 (64 IO tasks)

[Load balancing improvement]
15 x 360 x 720: use_io_procs increased from 15 to 60

[Old version of CalcStartandCount(), PIO_BUFFER_SIZE_LIMIT = 256 MB]
Case run time: 5 min 32 sec
"l:PIO:PIOc_createfile": 0.061 sec
"l:PIO:PIOc_put_att_tc": 0.072 sec
"l:PIO:PIOc_put_vars_tc": 0.963 sec
"l:PIO:PIOc_write_darray": 0.497 sec
"l:PIO:PIOc_closefile_write_mode": 230.777 sec
Total: 232.37 sec
Overall I/O write throughput: 1.512699165 GB / sec

[New version of CalcStartandCount(), PIO_BUFFER_SIZE_LIMIT = 256 MB]
Case run time: 4 min 5 sec
"l:PIO:PIOc_createfile": 0.093 sec
"l:PIO:PIOc_put_att_tc": 0.081 sec
"l:PIO:PIOc_put_vars_tc": 1.156 sec
"l:PIO:PIOc_write_darray": 0.527 sec
"l:PIO:PIOc_closefile_write_mode": 143.459 sec
Total: 145.316 sec
Overall I/O write throughput: 2.418907105 GB / sec